separated decoding logic from downloading correct pre downloading

+++ app/src/main/java/com/mmdev/kudago/app/core/utils/image_loader/cache/Cache.kt
@@ -53,5 +53,5 @@ interface Cache<Key : Any, Value : Any> {
 	 *
 	 * @return cache current size in bytes.
 	 */
-	fun size(): Long
+	fun size(): Int
 }
